+++ R/create_corpus.R
@@ -15,6 +15,25 @@
 #' @export
 create_corpus <- function(path){
 
+  #### test syntax of file names ####
+  filenames <- list.files(path)
+
+  tests <- filenames |>
+    sapply(\(x){
+
+      stringr::str_detect(x, "(.+)_(.+)\\.txt")
+
+    })
+
+  if(any(!tests)){
+
+    wrong.files <- names(tests[tests == FALSE]) |> paste(collapse = ", ")
+
+    stop("Some files do not follow the required syntax: ", wrong.files)
+
+  }
+
+  #### main function ####
   corpus <- readtext::readtext(
     file = paste0(path, "/*.txt"),
     docvarsfrom = "filenames", docvarnames = c("author", "textname")