+++ tests/testthat/test-create_corpus.R
@@ -5,3 +5,13 @@ test_that("corpus creation works", {
   expect_snapshot(create_corpus(path))
 
 })
+test_that("filename syntax check works", {
+
+  path <- testthat::test_path("data", "wrong_texts")
+
+  expect_error(
+    create_corpus(path),
+    "Some files do not follow the required syntax: abc2.txt, smithtext.txt"
+  )
+
+})