+++ R/create_corpus.R
@@ -3,6 +3,7 @@
 #' Function to read in text data and turn it into a `quanteda` corpus object.
 #'
 #' @param path A string containing the path to a folder of plain text files (ending in .txt) with their name structured as following: authorname_textname.txt (e.g. smith_text1.txt).
+#' @param encoding Either a single string indicating the encoding for all files or a vector of strings indicating the encodings for each file. The default is UTF-8.
 #'
 #' @return A `quanteda` corpus object with the authors' names and the text names as docvars.
 #'
@@ -13,7 +14,7 @@
 #' }
 #'
 #' @export
-create_corpus <- function(path) {
+create_corpus <- function(path, encoding = "UTF-8") {
   #### test syntax of file names ####
   filenames <- list.files(path)
 
@@ -30,7 +31,8 @@ create_corpus <- function(path) {
   #### main function ####
   corpus <- readtext::readtext(
     file = paste0(path, "/*.txt"),
-    docvarsfrom = "filenames", docvarnames = c("author", "textname")
+    docvarsfrom = "filenames", docvarnames = c("author", "textname"),
+    encoding = encoding
   ) |>
     quanteda::corpus()