+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * `create_corpus()` tests for the correct syntax of the file names and returns an error if not correct (plus showing which file names are incorrect).
 
+* `create_corpus()` includes an argument to specify the encoding of the texts.
+
 # idiolect 1.1.1
 
 * minor bug fixes