+++ NEWS.md
@@ -1,4 +1,4 @@
-# idiolect (development version)
+# idiolect 1.2.0
 
 * minor bug fixes